Notebook GPflow GPR on mesh
