+++ geometric_kernels/__init__.py
@@ -1,1 +1,3 @@
 import geometric_kernels._logging  # noqa: F401
+
+from geometric_kernels.lab_extras import *
