--- geometric_kernels/backends/pytorch/__init__.py
@@ -1,4 +1,0 @@
-"""
-GPyTorch and PyTorch specific code
-"""
-from geometric_kernels.backends.pytorch.gpytorch import GPytorchGeometricKernel
