--- geometric_kernels/backends/tensorflow/__init__.py
@@ -1,4 +1,0 @@
-"""
-GPflow and TensorFlow specific code
-"""
-from geometric_kernels.backends.tensorflow.gpflow import GPflowGeometricKernel
