+++ geometric_kernels/frontends/pytorch/__init__.py
@@ -1,0 +1,5 @@
+"""
+GPyTorch and PyTorch specific code
+"""
+import lab.torch
+import geometric_kernels.lab_extras.torch
