+++ geometric_kernels/frontends/tensorflow/__init__.py
@@ -1,0 +1,7 @@
+"""
+GPflow and TensorFlow specific code
+"""
+import lab.tensorflow
+import geometric_kernels.lab_extras.tensorflow
+
+print("frontends/tensorflow/__init__")
