+++ geometric_kernels/frontends/tensorflow/gpflow.py
@@ -11,6 +11,7 @@ from gpflow.utilities import positive
 from geometric_kernels.kernels import BaseGeometricKernel
 from geometric_kernels.spaces.base import Space
 
+print("frontends/tensorflow/gpflow.py")
 
 class GPflowGeometricKernel(gpflow.kernels.Kernel):
     """
@@ -39,3 +40,21 @@ class GPflowGeometricKernel(gpflow.kernels.Kernel):
     def K_diag(self, X):
         lengthscale = tf.convert_to_tensor(self.lengthscale)
         return self._kernel.K_diag(X, lengthscale=lengthscale)
+
+
+
+class DefaultFloatZeroMeanFunction(gpflow.mean_functions.Constant):
+    """
+    Zero mean function. The default GPflow `ZeroMeanFunction`
+    uses the input's dtype as output type, this minor adaptation
+    uses GPflow's `default_float` instead.
+    """
+
+    def __init__(self, output_dim=1):
+        super().__init__()
+        self.output_dim = output_dim
+        del self.c
+
+    def __call__(self, inputs):
+        output_shape = tf.concat([tf.shape(inputs)[:-1], [self.output_dim]], axis=0)
+        return tf.zeros(output_shape, dtype=gpflow.default_float())
