+++ geometric_kernels/lab_extras/__init__.py
@@ -1,3 +1,5 @@
 from lab import dispatch
 
 from geometric_kernels.lab_extras.extras import *
+# Always load the numpy backend because we assume numpy is always installed.
+from geometric_kernels.lab_extras.numpy import *
