+++ geometric_kernels/lab_extras/tensorflow/__init__.py
@@ -1,1 +1,3 @@
 from geometric_kernels.lab_extras.tensorflow.extras import *
+
+print("lab_extras/tensorflow/__init__.py")
